fix(files): render snekFinder even without initBackendLink
This change updates the rendering behaviour of the SnekFinder.
Before the loading state of the SnekFinder was always true, unless
a initBackendLink was set. Due to this, the rendering of the SnekFinder
failed for all applications that did not set the initBackendLink through
`jaen-pages.json` or cache.

Now the loading state is set to false regardless of the initBackendLink.
Without the initBackendLink,
```js
localStorage.removeItem('snek-finder-osg-backend-root')
```
is called to remove the cached snekFinder data.