feat(indexfield): add page creation with renderers
This change introduces the possibility to create a page through the
`IndexField`. This is useful for creating pages quickly and easily from
the index without having to open the `MainUI` and create a page within
the `PageExplorer`.

This concept features two new ways to create a page:
1. Use the `addPage` method of the `IndexField`.
2. Use the `defaultPopover` of the `IndexField`, that renders a so called
`renderer`.

Renderer:
Functions that accept `IndexRenderArgs` are called `renderers`. Those are
used to provide prebuild callback functions for the props `onRender` and
`onRenderPopver`.
The concept of renderes will also apply for other fields in the future.