feat: add templated email sending functionality (#20)
* feat: add props to ChevronRight icon in DayPicker

* WIP: add email page

In the future it should provide a way to send
templated emails from the website address.

It should also provide a way to preview the
templated email before sending it and a way to
schedule the email to be sent at a later time.

* feat: add templated email sending functionality

Features include:
 - Introduce the ability to select and send predefined email templates.
 - Implement a preview modal to view the filled email template before
   sending.
 - Integrate form validation using react-hook-form and Zod for robust
   input handling.
 - Add a rich text editor (Tiptap) for composing and editing email
   messages.
 - Enhance user experience with loading states and error handling for
   template fetching.
 - Ensure secure content rendering by sanitizing HTML with DOMPurify.

Some features like BCC and scheduling are not yet supported by the
backend and have therefore been disabled (css display none) in the UI.